cloud_function/requirements.txt
functions-framework==3.*
requests
google-cloud-storage
google-auth
